requirements.txt
app-version==0.2.1
BeautifulSoup==3.2.1
beautifulsoup4==4.4.1
Django==1.8.5
django-appconf==1.0.2
django-dajax==0.9.2
django-dajaxice==0.7
django-floppyforms==1.6.2
django-model-utils==2.5
django-notification==1.3.3
django-notifications==0.1.dev0
django-notifications-hq==1.2
django-user-accounts==1.3.1
fake-factory==0.2
html5lib==0.9999999
httplib2==0.9.2
ipaddress==1.0.16
jsonfield==1.0.3
notify==0.3.0
python-dateutil==2.5.3
pytz==2016.4
simplejson==3.8.2
six==1.10.0
tolerance==0.1.2
twilio==5.4.0
validate-email==1.3
MySQL-python
